Fix bugs in prepareNextTickTransactions()
Also log full tx storage case to console.

First bug was in not updating the tick transaction storage with pending
tx if there was already a version of the tx in the storage that didn't
match the digest.

The second bug was in counting numberOfKnownNextTickTransactions.
Tx are only known if we have both tx digest in tick data and the
matching tx in ts.tickTransactions.